+++ CalculatorProject/src/calculator/Lv2/MultiplyOperation.java
@@ -1,7 +1,7 @@
 package calculator.Lv2;
 
 public class MultiplyOperation {
-    public double operate(int firstNumber, int secondNumber){
+    public int operate(int firstNumber, int secondNumber){
         return firstNumber * secondNumber;
     }
 }
